refactoring powerplatform_publishers
