+++ .changes/unreleased/added-20260330-193000.yaml
@@ -1,5 +1,5 @@
 kind: added
-body: Add powerplatform_publisher resource and data source with Dataverse CRUD support and nested address handling
+body: Add powerplatform_publisher resource and powerplatform_publishers data source with Dataverse CRUD support and nested address handling
 time: 2026-03-30T19:30:00Z
 custom:
     Issue: 1133
