--- docs/data-sources/publisher.md
@@ -1,80 +1,0 @@
----
-# generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "powerplatform_publisher Data Source - Power Platform"
-subcategory: ""
-description: |-
-  Fetches a Dataverse publisher by Dataverse publisher id or unique name.
----
-
-# powerplatform_publisher (Data Source)
-
-Fetches a Dataverse publisher by Dataverse publisher id or unique name.
-
-## Example Usage
-
-```terraform
-data "powerplatform_publisher" "example" {
-  environment_id = "00000000-0000-0000-0000-000000000001"
-  uniquename     = "contoso"
-}
-```
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Required
-
-- `environment_id` (String) Id of the Dataverse-enabled environment containing the publisher.
-
-### Optional
-
-- `id` (String) Dataverse publisher id.
-- `timeouts` (Attributes) (see [below for nested schema](#nestedatt--timeouts))
-- `uniquename` (String) Unique name of the publisher.
-
-### Read-Only
-
-- `address` (Attributes List) Publisher addresses mapped from Dataverse address slots 1 and 2. (see [below for nested schema](#nestedatt--address))
-- `customization_option_value_prefix` (Number) Option value prefix used for option set values created by this publisher.
-- `customization_prefix` (String) Customization prefix used for solution components created by this publisher.
-- `description` (String) Description of the publisher.
-- `email_address` (String) Email address for the publisher.
-- `friendly_name` (String) Display name of the publisher.
-- `is_read_only` (Boolean) Whether Dataverse reports this publisher as read only.
-- `supporting_website_url` (String) Supporting website URL for the publisher.
-
-<a id="nestedatt--timeouts"></a>
-### Nested Schema for `timeouts`
-
-Optional:
-
-- `read` (String) A string that can be [parsed as a duration](https://pkg.go.dev/time#ParseDuration) consisting of numbers and unit suffixes, such as "30s" or "2h45m". Valid time units are "s" (seconds), "m" (minutes), "h" (hours). Read operations occur during any refresh or planning operation when refresh is enabled.
-
-
-<a id="nestedatt--address"></a>
-### Nested Schema for `address`
-
-Read-Only:
-
-- `address_id` (String) Dataverse address identifier for the slot.
-- `address_type_code` (Number) Address type code.
-- `city` (String) City name.
-- `country` (String) Country or region name.
-- `county` (String) County name.
-- `fax` (String) Fax number.
-- `latitude` (Number) Latitude value.
-- `line1` (String) Street line 1.
-- `line2` (String) Street line 2.
-- `line3` (String) Street line 3.
-- `longitude` (Number) Longitude value.
-- `name` (String) Address name.
-- `post_office_box` (String) Post office box.
-- `postal_code` (String) Postal code.
-- `shipping_method_code` (Number) Shipping method code.
-- `slot` (Number) Address slot number in Dataverse. Valid values are `1` and `2`.
-- `state_or_province` (String) State or province name.
-- `telephone1` (String) Primary telephone number.
-- `telephone2` (String) Secondary telephone number.
-- `telephone3` (String) Tertiary telephone number.
-- `ups_zone` (String) UPS zone value.
-- `utc_offset` (Number) UTC offset for the address.
