--- examples/data-sources/powerplatform_publisher/data-source.tf
@@ -1,4 +1,0 @@
-data "powerplatform_publisher" "example" {
-  environment_id = "00000000-0000-0000-0000-000000000001"
-  uniquename     = "contoso"
-}
