--- examples/data-sources/powerplatform_publisher/outputs.tf
@@ -1,3 +1,0 @@
-output "publisher" {
-  value = data.powerplatform_publisher.example
-}
