+++ examples/data-sources/powerplatform_publishers/data-source.tf
@@ -1,0 +1,15 @@
+terraform {
+  required_providers {
+    powerplatform = {
+      source = "microsoft/power-platform"
+    }
+  }
+}
+
+provider "powerplatform" {
+  use_cli = true
+}
+
+data "powerplatform_publishers" "example" {
+  environment_id = "e9f7a826-3dc2-e0ca-944e-630e3acead06"
+}
