+++ examples/data-sources/powerplatform_publishers/outputs.tf
@@ -1,0 +1,3 @@
+output "publishers" {
+  value = data.powerplatform_publishers.example.publishers
+}
